requirements.txt
streamlit==1.32.0
opencv-python==4.9.0.80
numpy==1.26.4
pandas==2.2.1
pillow==10.2.0
torch==2.2.1
torchvision==0.17.1
ultralytics==8.1.24
deep-sort-realtime==1.3.2
google-generativeai==0.4.0
tensorflow==2.16.1
tf-keras==2.16.0
protobuf==3.20.3
deepface==0.0.90
pyttsx3==2.90
